Add rootpage component
